fix int overflow issue

+++ dorado/read_pipeline/WriterNode.h
@@ -18,7 +18,7 @@ private:
     // Emit Fastq if true
     bool m_emit_fastq;
     // Total number of raw samples from the read WriterNode has processed. Used for performance benchmarking and debugging.
-    int m_num_samples_processed;
+    int64_t m_num_samples_processed;
     //Total number of reads WriterNode has processed
     int m_num_reads_processed;
     // Time when Node is initialised.
